moved plot bin locomotion to analyss, finalized plot_follow
