+++ src/analysis.py
@@ -1,11 +1,11 @@
 # File to analyse the results
 # Copied from Moritz Maxeiner Masterthesis: https://git.imp.fu-berlin.de/bioroboticslab/robofish/ai
 
-from numpy import float32
-from numpy.linalg import norm
 import reader
 import matplotlib.pyplot as plt
 from functions import *
+import pandas as pd
+import numpy as np
 
 
 def normalize_series(x):
@@ -13,7 +13,7 @@ def normalize_series(x):
     Given a series of vectors, return a series of normalized vectors.
     Null vectors are mapped to `NaN` vectors.
     """
-    return (x.T / norm(x, axis=-1)).T
+    return (x.T / np.linalg.norm(x, axis=-1)).T
 
 
 def calc_iid(a, b):
@@ -21,7 +21,7 @@ def calc_iid(a, b):
     Given two series of poses - with X and Y coordinates of their positions as the first two elements -
     return the inter-individual distance (between the positions).
     """
-    return norm(b[:, :2] - a[:, :2], axis=-1)
+    return np.linalg.norm(b[:, :2] - a[:, :2], axis=-1)
 
 
 def calc_tlvc(a, b, tau_min, tau_max):
@@ -30,7 +30,7 @@ def calc_tlvc(a, b, tau_min, tau_max):
     time lagged velocity correlation from the first to the second series.
     """
     length = tau_max - tau_min
-    return float32(
+    return np.float32(
         [
             (a[t] @ b[t + tau_min :][:length].T).mean()
             for t in range(min(len(a), len(b) - tau_max + 1))
@@ -79,6 +79,42 @@ def plot_follow(tracks, file = "data/follow.png", max_tolerated_movement=20):
     plt.bar(y_pos, valsToPlot)
     plt.xticks(y_pos, labels)
     plt.show()
+
+
+def plot_locomotion(paths, path_to_save, round):
+    """
+    paths should be iterable
+    path_to_save is the folder in which it will be saved
+    round will round to one decimal place (round = 2 -> all values will be of the form x.yz)
+    Creates Bar Plots for the Dataframes, one for linear_movement and for angle_radians
+    """
+    df_list = []
+    for path in paths:
+        df_list.append(pd.read_csv(path, sep = ";"))
+    df_all = pd.concat(df_list)
+    #round values
+    df_all = df_all.round(round)
+
+    #get all linear_movement columns respectivly all angle_radians columns
+    mov_cols = [col for col in df_all.columns if "linear_movement" in col]
+    ang_cols = [col for col in df_all.columns if "angle_radians" in col]
+
+    #melt linear_movement columns together in one
+    df_mov = df_all[mov_cols].melt(var_name = "columns", value_name = "value")
+    #count the values and plot it
+    ax = df_mov["value"].value_counts().sort_index().plot.bar()
+    for i, t in enumerate(ax.get_xticklabels()):
+        if (i % 2) != 0:
+            t.set_visible(False)
+    fig = ax.get_figure()
+    fig.set_size_inches(25, 12.5)
+    fig.savefig(path_to_save + "plot_linear_movement.png")
+
+    df_ang = df_all[ang_cols].melt(var_name = "columns", value_name = "value")
+    ax = df_ang["value"].value_counts().sort_index().plot.bar()
+    fig = ax.get_figure()
+    fig.set_size_inches(25, 12.5)
+    fig.savefig(path_to_save + "plot_angle_radians.png")
 
 
 def main():
